properties/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
CacheControl==0.12.6
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cssselect==1.1.0
distlib==0.3.0
distro==1.4.0
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
fake-headers==1.0.2
fake-useragent==0.1.11
html5lib==1.1
idna==2.10
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
lxml==4.6.3
msgpack==0.6.2
packaging==20.3
parse==1.19.0
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyee==8.1.0
PyJWT==2.1.0
pyparsing==2.4.6
pyppeteer==0.2.5
pyquery==1.4.3
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-html==0.10.0
retrying==1.3.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.60.0
urllib3==1.26.4
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
